Move to Core namespace
